Application for voucher

+++ Domain/Entities/Applications/Application.cs
@@ -10,13 +10,13 @@ namespace Domain.Entities.Applications;
 
 public class Application
 {
-    public long Id { get; set; }
+    public int Id { get; set; }
     public DateTime Date { get; set; }
     public int Number { get; set; }
     public bool StarterKit { get; set; }
     public PaymentMethodEnum Payment { get; set; }
     public int DistanceId { get; set; }
-    public int UserId { get; set; }
+    public long UserId { get; set; }
     public int MarathonId { get; set; }
     public int? PromocodeId { get; set; }
     public Promocode? Promocode { get; set; }
